feat: functional search with scrollback support
- find_matches() iterates full grid (history + viewport), case-insensitive
- SearchMatch.row is i32 — negative values index scrollback history
- scroll_to_current_match() adjusts display_offset to centre match in viewport
- build_match_set() returns HashSet<(usize, i32)> for renderer
- Fix scrollback rendering: viewport_row = raw_row + display_offset so
  history lines are visible when scrolled; cursor/selection/match coords
  all use raw_row for grid comparisons
- Match highlight: yellow-gold bg (#fac525) with dark text

+++ crates/SYNAPSE_-app/src/state.rs
@@ -15,7 +15,7 @@ pub struct DividerDrag {
 #[derive(Debug, Clone)]
 pub struct SearchMatch {
     pub col: usize,
-    pub row: usize,
+    pub row: i32,
 }
 
 pub struct SearchState {
